refactor(tests): move utility functions to helpers.py and update imports
